+++ target_chains/ethereum/examples/oracle_swap/app/src/OrderEntry.tsx
@@ -1,0 +1,206 @@
+import React, { useEffect, useState } from "react";
+import "./App.css";
+import Web3 from "web3";
+import { BigNumber } from "ethers";
+import { TokenConfig, numberToTokenQty, tokenQtyToNumber } from "./utils";
+import IPythAbi from "@pythnetwork/pyth-sdk-solidity/abis/IPyth.json";
+import OracleSwapAbi from "./abi/OracleSwapAbi.json";
+import { approveToken, getApprovedQuantity } from "./erc20";
+import { EvmPriceServiceConnection } from "@pythnetwork/pyth-evm-js";
+
+export function OrderEntry(props: {
+  web3: Web3 | undefined;
+  account: string | null;
+  isBuy: boolean;
+  approxPrice: number | undefined;
+  baseToken: TokenConfig;
+  quoteToken: TokenConfig;
+  priceServiceUrl: string;
+  pythContractAddress: string;
+  swapContractAddress: string;
+}) {
+  const [qty, setQty] = useState<string>("1");
+  const [qtyBn, setQtyBn] = useState<BigNumber | undefined>(
+    BigNumber.from("1")
+  );
+  const [authorizedQty, setAuthorizedQty] = useState<BigNumber>(
+    BigNumber.from("0")
+  );
+  const [isAuthorized, setIsAuthorized] = useState<boolean>(false);
+
+  const [spentToken, setSpentToken] = useState<TokenConfig>(props.baseToken);
+  const [approxQuoteSize, setApproxQuoteSize] = useState<number | undefined>(
+    undefined
+  );
+
+  useEffect(() => {
+    if (props.isBuy) {
+      setSpentToken(props.quoteToken);
+    } else {
+      setSpentToken(props.baseToken);
+    }
+  }, [props.isBuy]);
+
+  useEffect(() => {
+    async function helper() {
+      if (props.web3 !== undefined && props.account !== null) {
+        setAuthorizedQty(
+          await getApprovedQuantity(
+            props.web3!,
+            spentToken.erc20Address,
+            props.account!,
+            props.swapContractAddress
+          )
+        );
+      } else {
+        setAuthorizedQty(BigNumber.from("0"));
+      }
+    }
+
+    helper();
+  }, [props.web3, props.account, spentToken]);
+
+  useEffect(() => {
+    try {
+      const qtyBn = numberToTokenQty(qty, props.baseToken.decimals);
+      setQtyBn(qtyBn);
+    } catch (error) {
+      setQtyBn(undefined);
+    }
+  }, [qty]);
+
+  useEffect(() => {
+    if (qtyBn !== undefined) {
+      setIsAuthorized(authorizedQty.gte(qtyBn));
+    } else {
+      setIsAuthorized(false);
+    }
+  }, [qtyBn, authorizedQty]);
+
+  useEffect(() => {
+    if (qtyBn !== undefined && props.approxPrice !== undefined) {
+      setApproxQuoteSize(
+        tokenQtyToNumber(qtyBn, props.baseToken.decimals) * props.approxPrice
+      );
+    } else {
+      setApproxQuoteSize(undefined);
+    }
+  }, [props.approxPrice, qtyBn]);
+
+  return (
+    <div>
+      <div>
+        <p>
+          {props.isBuy ? "Buy" : "Sell"}
+          <input
+            type="text"
+            name="base"
+            value={qty}
+            onChange={(event) => {
+              setQty(event.target.value);
+            }}
+          />
+          {props.baseToken.name}
+        </p>
+        {qtyBn !== undefined && approxQuoteSize !== undefined ? (
+          props.isBuy ? (
+            <p>
+              Pay {approxQuoteSize.toFixed(3)} {props.quoteToken.name} to
+              receive{" "}
+              {tokenQtyToNumber(qtyBn, props.baseToken.decimals).toFixed(3)}{" "}
+              {props.baseToken.name}
+            </p>
+          ) : (
+            <p>
+              Pay {tokenQtyToNumber(qtyBn, props.baseToken.decimals).toFixed(3)}{" "}
+              {props.baseToken.name} to receive {approxQuoteSize.toFixed(3)}{" "}
+              {props.quoteToken.name}
+            </p>
+          )
+        ) : (
+          <p>Transaction details are loading...</p>
+        )}
+      </div>
+
+      <div className={"swap-steps"}>
+        {props.account === null || props.web3 === undefined ? (
+          <div>Connect your wallet to swap</div>
+        ) : (
+          <div>
+            1.{" "}
+            <button
+              onClick={async () => {
+                await approveToken(
+                  props.web3!,
+                  spentToken.erc20Address,
+                  props.account!,
+                  props.swapContractAddress
+                );
+              }}
+              disabled={isAuthorized}
+            >
+              {" "}
+              Approve{" "}
+            </button>
+            2.
+            <button
+              onClick={async () => {
+                await sendSwapTx(
+                  props.web3!,
+                  props.priceServiceUrl,
+                  props.baseToken.pythPriceFeedId,
+                  props.quoteToken.pythPriceFeedId,
+                  props.pythContractAddress,
+                  props.swapContractAddress,
+                  props.account!,
+                  qtyBn!,
+                  props.isBuy
+                );
+              }}
+              disabled={!isAuthorized}
+            >
+              {" "}
+              Submit{" "}
+            </button>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+}
+
+async function sendSwapTx(
+  web3: Web3,
+  priceServiceUrl: string,
+  baseTokenPriceFeedId: string,
+  quoteTokenPriceFeedId: string,
+  pythContractAddress: string,
+  swapContractAddress: string,
+  sender: string,
+  qtyWei: BigNumber,
+  isBuy: boolean
+) {
+  const pythPriceService = new EvmPriceServiceConnection(priceServiceUrl);
+  const priceFeedUpdateData = await pythPriceService.getPriceFeedsUpdateData([
+    baseTokenPriceFeedId,
+    quoteTokenPriceFeedId,
+  ]);
+
+  const pythContract = new web3.eth.Contract(
+    IPythAbi as any,
+    pythContractAddress
+  );
+
+  const updateFee = await pythContract.methods
+    .getUpdateFee(priceFeedUpdateData.length)
+    .call();
+
+  const swapContract = new web3.eth.Contract(
+    OracleSwapAbi as any,
+    swapContractAddress
+  );
+
+  await swapContract.methods
+    .swap(isBuy, qtyWei, priceFeedUpdateData)
+    .send({ value: updateFee, from: sender });
+}
